+++ src/main/java/dk/cngroup/lentils/controller/TeamController.java
@@ -1,7 +1,7 @@
 package dk.cngroup.lentils.controller;
 
+import dk.cngroup.lentils.dto.TeamFormDTO;
 import dk.cngroup.lentils.entity.Team;
-import dk.cngroup.lentils.entity.formEntity.TeamDTO;
 import dk.cngroup.lentils.service.TeamService;
 import dk.cngroup.lentils.service.convertors.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,19 +37,19 @@ public class TeamController {
 
     @GetMapping
     public String addTeam(final Model model) {
-        fillModelAttributes(model, teamService.getAll(), new TeamDTO(), ACTION_TEAM_SAVE);
+        fillModelAttributes(model, teamService.getAll(), new TeamFormDTO(), ACTION_TEAM_SAVE);
         return VIEW_PATH;
     }
 
     @PostMapping(value = "/add")
-    public String addTeam(@Valid @ModelAttribute("team") final TeamDTO teamDto,
+    public String addTeam(@Valid @ModelAttribute("team") final TeamFormDTO teamFormDto,
             final BindingResult bindingResult,
             final Model model) {
         Team team = new Team();
-        mapper.map(teamDto, team);
+        mapper.map(teamFormDto, team);
         teamService.checkUsernameIsUnique(team, bindingResult);
         if (bindingResult.hasErrors()) {
-            fillModelAttributes(model, teamService.getAll(), teamDto, ACTION_TEAM_SAVE);
+            fillModelAttributes(model, teamService.getAll(), teamFormDto, ACTION_TEAM_SAVE);
             return VIEW_PATH;
         }
         teamService.save(team);
@@ -60,21 +60,21 @@ public class TeamController {
     public String update(@PathVariable("id") final Long id, final Model model) {
         fillModelAttributes(model,
                 teamService.getAll(),
-                mapper.map(teamService.getTeam(id), TeamDTO.class),
+                mapper.map(teamService.getTeam(id), TeamFormDTO.class),
                 ACTION_TEAM_UPDATE + id);
         return VIEW_PATH;
     }
 
     @PostMapping("/update/{id}")
     public String update(@PathVariable("id") final Long id,
-            @Valid @ModelAttribute("team") final TeamDTO teamDto,
+            @Valid @ModelAttribute("team") final TeamFormDTO teamFormDto,
             final BindingResult bindingResult,
             final Model model) {
         Team team = teamService.getTeam(id);
-        mapper.map(teamDto, team);
+        mapper.map(teamFormDto, team);
         teamService.checkUsernameIsUnique(team, bindingResult);
         if (bindingResult.hasErrors()) {
-            fillModelAttributes(model, teamService.getAll(), teamDto, ACTION_TEAM_UPDATE + id);
+            fillModelAttributes(model, teamService.getAll(), teamFormDto, ACTION_TEAM_UPDATE + id);
             return VIEW_PATH;
         }
         teamService.update(team);
@@ -90,11 +90,11 @@ public class TeamController {
     private void fillModelAttributes(
             final Model model,
             final List<Team> teams,
-            final TeamDTO teamDto,
+            final TeamFormDTO teamFormDto,
             final String action
     ) {
         model.addAttribute("teams", teams);
-        model.addAttribute("teamDto", teamDto);
+        model.addAttribute("teamFormDto", teamFormDto);
         model.addAttribute("action", action);
     }
 }
