+++ src/main/resources/templates/team/main.html
@@ -8,7 +8,7 @@
     <div class="row">
         <div class="col-sm-6">
             <h2>Tým</h2>
-            <form action="#" th:action="${action}" th:object="${teamDto}" method="post">
+            <form action="#" th:action="${action}" th:object="${teamFormDto}" method="post">
                 <div class="form-group">
                     <label for="name" class="input-value-required">Jméno</label>
                     <input id="name" class="form-control" type="text" th:field="*{name}" required/>
